Deleted browse event and updated explore event class. Explore event is meant to be for the user and all events is for the admin
