+++ src/app/src/main/java/com/example/qrcheckin/user/exploreevent/ExploreEventViewModel.java
@@ -1,11 +1,18 @@
-package com.example.qrcheckin.user.exploreEvent;
+package com.example.qrcheckin.user.exploreevent;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
+
+/**
+ * Creates a specialized ViewModel for the explore event fragment
+ */
 public class ExploreEventViewModel extends ViewModel {
     private final MutableLiveData<String> mText;
 
+    /**
+     * Initializes the ExploreEventViewModel class
+     */
     public ExploreEventViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is exploreEvent fragment");
@@ -13,4 +20,3 @@ public class ExploreEventViewModel extends ViewModel {
 
     public LiveData<String> getText() { return mText; }
 }
-
