+++ src/app/src/main/res/layout/fragment_explore_event.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".user.exploreEvent.ExploreEventFragment">
+    tools:context=".user.exploreevent.ExploreEventFragment">
     <TextView
         android:id="@+id/text_explore_event"
         android:layout_width="match_parent"
@@ -19,11 +19,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/explore_event_list_view"
-        app:layout_constraintTop_toTopOf="parent" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
+
