+++ src/app/src/main/res/navigation/mobile_navigation.xml
@@ -33,15 +33,9 @@
 
     <fragment
         android:id="@+id/nav_explore_event"
-        android:name="com.example.qrcheckin.user.browseevent.BrowseEventFragment"
+        android:name="com.example.qrcheckin.user.exploreevent.ExploreEventFragment"
         android:label="@string/nav_explore_events"
-        tools:layout="@layout/fragment_browse_events" />
-
-    <fragment
-        android:id="@+id/nav_explore_event"
-        android:name="com.example.qrcheckin.user.browseevent.BrowseEventFragment"
-        android:label="@string/nav_explore_events"
-        tools:layout="@layout/fragment_browse_events" />
+        tools:layout="@layout/fragment_explore_event" />
 
     <fragment
         android:id="@+id/nav_profile"
